@@ -1,7 +1,9 @@
 import 'package:asuka/asuka.dart';
 import 'package:clash_for_flutter/app/component/sys_app_bar.dart';
 import 'package:clash_for_flutter/app/exceptions/message_exception.dart';
-import 'package:clash_for_flutter/app/source/global_config.dart';
+import 'package:clash_for_flutter/app/source/app_config.dart';
+import 'package:clash_for_flutter/app/source/core_config.dart';
+import 'package:clash_for_flutter/app/utils/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
@@ -16,21 +18,25 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  final GlobalConfig _config = Modular.get<GlobalConfig>();
+  final AppConfig _config = Modular.get<AppConfig>();
+  final CoreConfig _core = Modular.get<CoreConfig>();
   bool _loading = false;
-
-  @override
-  void initState() {
-    super.initState();
-  }
 
   click() {
     setState(() => _loading = true);
     Future(() {
-      if (_config.systemProxy) {
-        return _config.closeProxy();
+      if (_config.tunIf) {
+        if (_core.tunEnable) {
+          return _core.closeTun();
+        } else {
+          return _core.openTun();
+        }
       } else {
-        return _config.openProxy();
+        if (_config.systemProxy) {
+          return _config.closeProxy();
+        } else {
+          return _config.openProxy();
+        }
       }
     }).catchError((e) {
       if (e is MessageException) {
@@ -41,22 +47,59 @@ class _HomePageState extends State<HomePage> {
     }).then((_) => setState(() => _loading = false));
   }
 
+  changeTun(bool? tunIf) {
+    if (tunIf ?? false) {
+      Asuka.showSnackBar(const SnackBar(content: Text("请注意 tun 模式需要以管理员运行该软件，否则将无法启用代理")));
+    }
+    _config.setState(tunIf: tunIf);
+  }
+
+  get idOpen {
+    if (_config.tunIf) {
+      return _core.tunEnable;
+    } else {
+      return _config.systemProxy;
+    }
+  }
+
+  Widget get _openButton {
+    if (_loading) {
+      return const LoadingButton();
+    }
+    return Observer(
+      builder: (_) {
+        if (idOpen) {
+          return OffButton(onTap: click);
+        } else {
+          return OnButton(onTap: click);
+        }
+      },
+    );
+  }
+
+  Widget get _switch {
+    if (Constants.isDesktop) {
+      return Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          const Text("Tun模式:"),
+          Observer(builder: (_) => Switch(value: _config.tunIf, onChanged: changeTun)),
+        ],
+      );
+    }
+    return Container();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: const SysAppBar(title: Text("Clash For Flutter")),
-      body: Center(
-        child: _loading
-            ? const LoadingButton()
-            : Observer(
-                builder: (_) {
-                  if (_config.systemProxy) {
-                    return OffButton(onTap: click);
-                  } else {
-                    return OnButton(onTap: click);
-                  }
-                },
-              ),
+      appBar: const SysAppBar(title: Text("Clash for Flutter")),
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          _openButton,
+          _switch,
+        ],
       ),
     );
   }
@@ -65,7 +108,7 @@ class _HomePageState extends State<HomePage> {
 class OnButton extends StatelessWidget {
   final Function() onTap;
 
-  const OnButton({Key? key, required this.onTap}) : super(key: key);
+  const OnButton({super.key, required this.onTap});
 
   @override
   Widget build(BuildContext context) {
@@ -103,7 +146,7 @@ class OnButton extends StatelessWidget {
 class OffButton extends StatelessWidget {
   final Function() onTap;
 
-  const OffButton({Key? key, required this.onTap}) : super(key: key);
+  const OffButton({super.key, required this.onTap});
 
   @override
   Widget build(BuildContext context) {
@@ -139,7 +182,7 @@ class OffButton extends StatelessWidget {
 }
 
 class LoadingButton extends StatelessWidget {
-  const LoadingButton({Key? key}) : super(key: key);
+  const LoadingButton({super.key});
 
   @override
   Widget build(BuildContext context) {
