@@ -5,8 +5,7 @@ import 'package:dart_json_mapper/dart_json_mapper.dart';
 enum GroupType { Selector, URLTest, Fallback, LoadBalance }
 
 extension GroupTypeValue on GroupType {
-  static List<String> get valueList =>
-      GroupType.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => GroupType.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => GroupTypeValue.valueList[index];
 }
@@ -15,8 +14,7 @@ extension GroupTypeValue on GroupType {
 enum VehicleType { HTTP, File, Compatible }
 
 extension VehicleTypeValue on VehicleType {
-  static List<String> get valueList =>
-      VehicleType.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => VehicleType.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => VehicleTypeValue.valueList[index];
 }
@@ -26,8 +24,7 @@ extension VehicleTypeValue on VehicleType {
 enum UsedProxy { DIRECT, REJECT, GLOBAL }
 
 extension UsedProxyValue on UsedProxy {
-  static List<String> get valueList =>
-      UsedProxy.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => UsedProxy.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => UsedProxyValue.valueList[index];
 }
@@ -36,8 +33,7 @@ extension UsedProxyValue on UsedProxy {
 enum Mode { Rule, Global, Direct }
 
 extension ModeValue on Mode {
-  static List<String> get valueList =>
-      Mode.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => Mode.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => ModeValue.valueList[index];
 }
@@ -45,8 +41,7 @@ extension ModeValue on Mode {
 enum DataUnit { Byte, KB, MB, GB, TB, PB }
 
 extension DataUnitValue on DataUnit {
-  static List<String> get valueList =>
-      DataUnit.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => DataUnit.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => DataUnitValue.valueList[index];
 }
@@ -65,19 +60,17 @@ extension SortTypeShowName on SortType {
 enum ProfileType { URL, FILE }
 
 extension ProfileTypeValue on ProfileType {
-  static List<String> get valueList =>
-      ProfileType.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => ProfileType.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => ProfileTypeValue.valueList[index];
 }
 
 /// 日志等级
 @jsonSerializable
-enum LogLevel { info, warning, error, debug, silent }
+enum LogLevel { debug, info, warning, error, silent }
 
 extension LogLevelValue on LogLevel {
-  static List<String> get valueList =>
-      LogLevel.values.map((e) => e.toString().split(".").last).toList();
+  static List<String> get valueList => LogLevel.values.map((e) => e.toString().split(".").last).toList();
 
   String get value => LogLevelValue.valueList[index];
 }
