@@ -1,7 +1,4 @@
 import 'package:clash_for_flutter/app/bean/group_bean.dart';
-import 'package:clash_for_flutter/app/bean/provider_bean.dart';
-import 'package:clash_for_flutter/app/bean/proxy_bean.dart';
-import 'package:clash_for_flutter/app/bean/proxie_show_bean.dart';
 import 'package:clash_for_flutter/app/enum/type_enum.dart';
 import 'package:mobx/mobx.dart';
 
@@ -21,26 +18,18 @@ abstract class ProxysModelBase with Store {
 
   /// 代理列表
   @observable
-  var proxies = <Proxy>[];
-  @observable
-  var providers = <String, Provider>{};
-  @observable
-  var proxiesMap = <String, List<ProxieShow>>{};
+  Map<String, dynamic> proxiesMap = {};
 
   @action
   setState({
     Group? global,
     SortType? sortType,
     List<Group>? groups,
-    List<Proxy>? proxies,
-    Map<String, Provider>? providers,
-    Map<String, List<ProxieShow>>? proxiesMap,
+    Map<String, dynamic>? proxiesMap,
   }) {
     if (global != null) this.global = global;
     if (sortType != null) this.sortType = sortType;
     if (groups != null) this.groups = groups;
-    if (proxies != null) this.proxies = proxies;
-    if (providers != null) this.providers = providers;
     if (proxiesMap != null) this.proxiesMap = proxiesMap;
   }
 }
