@@ -1,17 +1,17 @@
-import 'package:flutter_modular/flutter_modular.dart';
 import 'package:clash_for_flutter/app/pages/proxys/model/proxys_model.dart';
 import 'package:clash_for_flutter/app/pages/proxys/proxys_controller.dart';
 import 'package:clash_for_flutter/app/pages/proxys/proxys_page.dart';
+import 'package:flutter_modular/flutter_modular.dart';
 
 class ProxysModule extends Module {
   @override
-  final List<Bind> binds = [
-    Bind.singleton((_) => ProxysModel()),
-    Bind.factory((_) => ProxysController()),
-  ];
+  void binds(i) {
+    i.addSingleton(ProxysModel.new);
+    i.add(ProxysController.new);
+  }
 
   @override
-  final List<ModularRoute> routes = [
-    ChildRoute("/", child: (_, __) => const ProxysPage()),
-  ];
+  void routes(r) {
+    r.child("/", child: (_) => const ProxysPage());
+  }
 }
