@@ -4,6 +4,7 @@ import 'package:clash_for_flutter/app/component/sys_app_bar.dart';
 import 'package:clash_for_flutter/app/enum/type_enum.dart';
 import 'package:clash_for_flutter/app/pages/proxys/proxys_controller.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
@@ -14,21 +15,41 @@ class ProxysPage extends StatefulWidget {
   const ProxysPage({super.key});
 
   @override
-  ModularState<ProxysPage, ProxysController> createState() => _ProxysPageState();
+  State<ProxysPage> createState() => _ProxysPageState();
 }
 
-class _ProxysPageState extends ModularState<ProxysPage, ProxysController> {
+class _ProxysPageState extends State<ProxysPage> {
+  final ScrollController _scrollController = ScrollController();
+  final ProxysController _controller = Modular.get<ProxysController>();
+  bool _showFab = true;
+
   @override
   void initState() {
     super.initState();
-    controller.initState();
+    _controller.initState();
+    _scrollController.addListener(_scrollListener);
+  }
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
+  }
+
+  void _scrollListener() {
+    if (_scrollController.position.userScrollDirection == ScrollDirection.reverse && _showFab) {
+      setState(() => _showFab = false);
+    }
+    if (_scrollController.position.userScrollDirection == ScrollDirection.forward && !_showFab) {
+      setState(() => _showFab = true);
+    }
   }
 
   void testDelay(TabController tabController) async {
     var overlay = Loading.builder();
     Asuka.addOverlay(overlay);
-    await controller.delayGroup(
-      controller.model.groups[tabController.index],
+    await _controller.delayGroup(
+      _controller.model.groups[tabController.index],
     );
     overlay.remove();
   }
@@ -44,7 +65,7 @@ class _ProxysPageState extends ModularState<ProxysPage, ProxysController> {
 
   sortAction() {
     change(sortType, BuildContext context) {
-      controller.sortProxies(type: sortType);
+      _controller.sort(sortType);
       Navigator.of(context).pop();
     }
 
@@ -67,7 +88,7 @@ class _ProxysPageState extends ModularState<ProxysPage, ProxysController> {
                 title: Text(SortType.values[i].showName),
                 trailing: Radio<SortType>(
                   value: SortType.values[i],
-                  groupValue: controller.model.sortType,
+                  groupValue: _controller.model.sortType,
                   onChanged: (v) => change(v, cxt),
                 ),
               );
@@ -81,8 +102,7 @@ class _ProxysPageState extends ModularState<ProxysPage, ProxysController> {
   @override
   Widget build(BuildContext context) {
     return Observer(builder: (c) {
-      var groups = controller.model.groups;
-      var proxiesMap = controller.model.proxiesMap;
+      var groups = _controller.model.groups;
       return DefaultTabController(
         length: groups.length,
         child: Scaffold(
@@ -100,32 +120,6 @@ class _ProxysPageState extends ModularState<ProxysPage, ProxysController> {
                 icon: const Icon(Icons.sort_outlined),
                 onPressed: sortAction,
               ),
-/*
-              PopupMenuButton(
-                onSelected: (MenuType type) => moreMenu(type),
-                itemBuilder: (_) => <PopupMenuEntry<MenuType>>[
-                  PopupMenuItem(
-                    padding: const EdgeInsets.symmetric(horizontal: 16),
-                    value: MenuType.Sort,
-                    child: SizedBox(
-                      width: 100,
-                      child: Row(children: [
-                        Icon(
-                          Icons.sort_outlined,
-                          color: DefaultTextStyle.of(context).style.color,
-                        ),
-                        Expanded(
-                          child: Container(
-                            margin: const EdgeInsets.only(left: 10),
-                            child: const Text("排序"),
-                          ),
-                        ),
-                      ]),
-                    ),
-                  ),
-                ],
-              ),
-*/
             ],
           ),
           body: groups.isNotEmpty
@@ -133,63 +127,55 @@ class _ProxysPageState extends ModularState<ProxysPage, ProxysController> {
                   children: groups.map((group) {
                     var groupName = group.name;
                     var groupNow = group.now;
-                    var proxies = proxiesMap[groupName] ?? [];
+                    var list = _controller.getShowList(group);
                     return ListView.separated(
-                      separatorBuilder: (_, __) => const Divider(height: 5),
-                      itemCount: proxies.length,
+                      controller: _scrollController,
                       itemBuilder: (_, i) {
-                        var proxie = proxies[i];
-                        var proxieName = proxie.name;
-                        var delay = proxie.delay < 0
-                            ? null
-                            : Text(
-                                proxie.delay == 0 ? "..." : proxie.delay.toString(),
-                              );
-                        var subTitle = StringBuffer();
-                        if (proxie.type != null) {
-                          subTitle.write(proxie.type);
-                        }
-                        if (proxie.now != null) {
-                          subTitle.write(" [${proxie.now}]");
-                        }
+                        var show = list[i];
+                        var name = show.name;
+                        var delay = show.delay < 0 ? null : Text(show.delay == 0 ? "..." : show.delay.toString());
                         return ListTile(
                           visualDensity: const VisualDensity(
                             vertical: VisualDensity.minimumDensity,
                           ),
-                          selected: groupNow == proxieName,
+                          selected: groupNow == name,
                           title: Text(
-                            proxieName,
+                            name,
                             style: const TextStyle(fontSize: 14),
                           ),
                           subtitle: Text(
-                            subTitle.toString(),
+                            show.subTitle,
                             style: const TextStyle(fontSize: 12),
                           ),
-                          onTap: () => controller.select(
-                            name: groupName,
-                            select: proxieName,
-                          ),
                           trailing: delay,
+                          onTap: () => _controller.select(
+                            name: groupName,
+                            select: name,
+                          ),
                         );
                       },
+                      separatorBuilder: (_, __) => const Divider(height: 5),
+                      itemCount: list.length,
                     );
                   }).toList(),
                 )
               : const Center(child: Text("暂无可选代理节点")),
-          floatingActionButton: Builder(
-            builder: (cxt) {
-              var tabController = DefaultTabController.of(cxt);
-              return FloatingActionButton(
-                tooltip: "测延迟",
-                onPressed: () {
-                  if (groups.isNotEmpty) {
-                    testDelay(tabController!);
-                  }
-                },
-                child: const Icon(Icons.flash_on),
-              );
-            },
-          ),
+          floatingActionButton: _showFab
+              ? Builder(
+                  builder: (cxt) {
+                    var tabController = DefaultTabController.of(cxt);
+                    return FloatingActionButton(
+                      tooltip: "测延迟",
+                      onPressed: () {
+                        if (groups.isNotEmpty) {
+                          testDelay(tabController);
+                        }
+                      },
+                      child: const Icon(Icons.flash_on),
+                    );
+                  },
+                )
+              : null,
         ),
       );
     });
