@@ -1,6 +1,7 @@
 // AUTO GENERATED FILE, DO NOT EDIT.
 //
 // Generated by `package:ffigen`.
+// ignore_for_file: type=lint
 import 'dart:ffi' as ffi;
 
 class Clash {
@@ -50,7 +51,7 @@ class Clash {
   }
 
   late final ___security_check_cookiePtr =
-      _lookup<ffi.NativeFunction<ffi.Void Function(uintptr_t)>>(
+      _lookup<ffi.NativeFunction<ffi.Void Function(ffi.UintPtr)>>(
           '__security_check_cookie');
   late final ___security_check_cookie =
       ___security_check_cookiePtr.asFunction<void Function(int)>();
@@ -64,13 +65,13 @@ class Clash {
   }
 
   late final ___report_gsfailurePtr =
-      _lookup<ffi.NativeFunction<ffi.Void Function(uintptr_t)>>(
+      _lookup<ffi.NativeFunction<ffi.Void Function(ffi.UintPtr)>>(
           '__report_gsfailure');
   late final ___report_gsfailure =
       ___report_gsfailurePtr.asFunction<void Function(int)>();
 
-  late final ffi.Pointer<uintptr_t> ___security_cookie =
-      _lookup<uintptr_t>('__security_cookie');
+  late final ffi.Pointer<ffi.UintPtr> ___security_cookie =
+      _lookup<ffi.UintPtr>('__security_cookie');
 
   int get __security_cookie => ___security_cookie.value;
 
@@ -119,7 +120,7 @@ class Clash {
               ffi.Pointer<ffi.WChar>,
               ffi.Pointer<ffi.WChar>,
               ffi.UnsignedInt,
-              uintptr_t)>>('_invoke_watson');
+              ffi.UintPtr)>>('_invoke_watson');
   late final __invoke_watson = __invoke_watsonPtr.asFunction<
       void Function(ffi.Pointer<ffi.WChar>, ffi.Pointer<ffi.WChar>,
           ffi.Pointer<ffi.WChar>, int, int)>();
@@ -171,7 +172,7 @@ class Clash {
   }
 
   late final ___threadhandlePtr =
-      _lookup<ffi.NativeFunction<uintptr_t Function()>>('__threadhandle');
+      _lookup<ffi.NativeFunction<ffi.UintPtr Function()>>('__threadhandle');
   late final ___threadhandle = ___threadhandlePtr.asFunction<int Function()>();
 
   double cabs(
@@ -854,89 +855,74 @@ class Clash {
   late final __FCmulcr =
       __FCmulcrPtr.asFunction<_Fcomplex Function(_Fcomplex, double)>();
 
-  int setHomeDir(
+  int SetHomeDir(
     ffi.Pointer<ffi.Char> homeStr,
   ) {
-    return _setHomeDir(
+    return _SetHomeDir(
       homeStr,
     );
   }
 
-  late final _setHomeDirPtr =
+  late final _SetHomeDirPtr =
       _lookup<ffi.NativeFunction<GoUint8 Function(ffi.Pointer<ffi.Char>)>>(
-          'setHomeDir');
-  late final _setHomeDir =
-      _setHomeDirPtr.asFunction<int Function(ffi.Pointer<ffi.Char>)>();
+          'SetHomeDir');
+  late final _SetHomeDir =
+      _SetHomeDirPtr.asFunction<int Function(ffi.Pointer<ffi.Char>)>();
 
-  int setConfig(
+  int SetConfig(
     ffi.Pointer<ffi.Char> configStr,
   ) {
-    return _setConfig(
+    return _SetConfig(
       configStr,
     );
   }
 
-  late final _setConfigPtr =
+  late final _SetConfigPtr =
       _lookup<ffi.NativeFunction<GoUint8 Function(ffi.Pointer<ffi.Char>)>>(
-          'setConfig');
-  late final _setConfig =
-      _setConfigPtr.asFunction<int Function(ffi.Pointer<ffi.Char>)>();
+          'SetConfig');
+  late final _SetConfig =
+      _SetConfigPtr.asFunction<int Function(ffi.Pointer<ffi.Char>)>();
 
-  void withExternalController(
-    ffi.Pointer<ffi.Char> externalController,
-  ) {
-    return _withExternalController(
-      externalController,
-    );
-  }
-
-  late final _withExternalControllerPtr =
-      _lookup<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Char>)>>(
-          'withExternalController');
-  late final _withExternalController = _withExternalControllerPtr
-      .asFunction<void Function(ffi.Pointer<ffi.Char>)>();
-
-  void withSecret(
-    ffi.Pointer<ffi.Char> secret,
-  ) {
-    return _withSecret(
-      secret,
-    );
-  }
-
-  late final _withSecretPtr =
-      _lookup<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Char>)>>(
-          'withSecret');
-  late final _withSecret =
-      _withSecretPtr.asFunction<void Function(ffi.Pointer<ffi.Char>)>();
-
-  int mmdbVerify(
+  int VerifyMMDB(
     ffi.Pointer<ffi.Char> path,
   ) {
-    return _mmdbVerify(
+    return _VerifyMMDB(
       path,
     );
   }
 
-  late final _mmdbVerifyPtr =
+  late final _VerifyMMDBPtr =
       _lookup<ffi.NativeFunction<GoUint8 Function(ffi.Pointer<ffi.Char>)>>(
-          'mmdbVerify');
-  late final _mmdbVerify =
-      _mmdbVerifyPtr.asFunction<int Function(ffi.Pointer<ffi.Char>)>();
+          'VerifyMMDB');
+  late final _VerifyMMDB =
+      _VerifyMMDBPtr.asFunction<int Function(ffi.Pointer<ffi.Char>)>();
 
-  int startService() {
-    return _startService();
+  ffi.Pointer<ffi.Char> StartRust(
+    ffi.Pointer<ffi.Char> addr,
+  ) {
+    return _StartRust(
+      addr,
+    );
   }
 
-  late final _startServicePtr =
-      _lookup<ffi.NativeFunction<GoUint8 Function()>>('startService');
-  late final _startService = _startServicePtr.asFunction<int Function()>();
+  late final _StartRustPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Pointer<ffi.Char> Function(ffi.Pointer<ffi.Char>)>>('StartRust');
+  late final _StartRust = _StartRustPtr.asFunction<
+      ffi.Pointer<ffi.Char> Function(ffi.Pointer<ffi.Char>)>();
+
+  int StartService() {
+    return _StartService();
+  }
+
+  late final _StartServicePtr =
+      _lookup<ffi.NativeFunction<GoUint8 Function()>>('StartService');
+  late final _StartService = _StartServicePtr.asFunction<int Function()>();
 }
 
 typedef va_list = ffi.Pointer<ffi.Char>;
-typedef uintptr_t = ffi.UnsignedLongLong;
 
-class __crt_locale_data_public extends ffi.Struct {
+final class __crt_locale_data_public extends ffi.Struct {
   external ffi.Pointer<ffi.UnsignedShort> _locale_pctype;
 
   @ffi.Int()
@@ -946,17 +932,17 @@ class __crt_locale_data_public extends ffi.Struct {
   external int _locale_lc_codepage;
 }
 
-class __crt_locale_pointers extends ffi.Struct {
+final class __crt_locale_pointers extends ffi.Struct {
   external ffi.Pointer<__crt_locale_data> locinfo;
 
   external ffi.Pointer<__crt_multibyte_data> mbcinfo;
 }
 
-class __crt_locale_data extends ffi.Opaque {}
+final class __crt_locale_data extends ffi.Opaque {}
 
-class __crt_multibyte_data extends ffi.Opaque {}
+final class __crt_multibyte_data extends ffi.Opaque {}
 
-class _Mbstatet extends ffi.Struct {
+final class _Mbstatet extends ffi.Struct {
   @ffi.UnsignedLong()
   external int _Wchar;
 
@@ -968,8 +954,9 @@ class _Mbstatet extends ffi.Struct {
 }
 
 typedef errno_t = ffi.Int;
+typedef Darterrno_t = int;
 
-class _GoString_ extends ffi.Struct {
+final class _GoString_ extends ffi.Struct {
   external ffi.Pointer<ffi.Char> p;
 
   @ptrdiff_t()
@@ -977,29 +964,30 @@ class _GoString_ extends ffi.Struct {
 }
 
 typedef ptrdiff_t = ffi.LongLong;
+typedef Dartptrdiff_t = int;
 
-class _C_double_complex extends ffi.Struct {
+final class _C_double_complex extends ffi.Struct {
   @ffi.Array.multi([2])
   external ffi.Array<ffi.Double> _Val;
 }
 
-class _C_float_complex extends ffi.Struct {
+final class _C_float_complex extends ffi.Struct {
   @ffi.Array.multi([2])
   external ffi.Array<ffi.Float> _Val;
 }
 
-class _C_ldouble_complex extends ffi.Opaque {}
+final class _C_ldouble_complex extends ffi.Opaque {}
 
 typedef _Dcomplex = _C_double_complex;
 typedef _Fcomplex = _C_float_complex;
 
-class GoInterface extends ffi.Struct {
+final class GoInterface extends ffi.Struct {
   external ffi.Pointer<ffi.Void> t;
 
   external ffi.Pointer<ffi.Void> v;
 }
 
-class GoSlice extends ffi.Struct {
+final class GoSlice extends ffi.Struct {
   external ffi.Pointer<ffi.Void> data;
 
   @GoInt()
@@ -1011,7 +999,9 @@ class GoSlice extends ffi.Struct {
 
 typedef GoInt = GoInt64;
 typedef GoInt64 = ffi.LongLong;
+typedef DartGoInt64 = int;
 typedef GoUint8 = ffi.UnsignedChar;
+typedef DartGoUint8 = int;
 
 const int _VCRT_COMPILER_PREPROCESSOR = 1;
 
@@ -1041,6 +1031,10 @@ const int _HAS_CXX23 = 0;
 
 const int _HAS_NODISCARD = 1;
 
+const int _ARM_WINAPI_PARTITION_DESKTOP_SDK_AVAILABLE = 1;
+
+const int _CRT_BUILD_DESKTOP_APP = 1;
+
 const int _UCRT_DISABLED_WARNINGS = 4324;
 
 const int _ARGMAX = 100;
@@ -1057,9 +1051,7 @@ const int _CRT_FUNCTIONS_REQUIRED = 1;
 
 const int _CRT_HAS_CXX17 = 0;
 
-const int _ARM_WINAPI_PARTITION_DESKTOP_SDK_AVAILABLE = 1;
-
-const int _CRT_BUILD_DESKTOP_APP = 1;
+const int _CRT_HAS_C11 = 0;
 
 const int _CRT_INTERNAL_NONSTDC_NAMES = 1;
 
