@@ -6,14 +6,20 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <desktop_lifecycle/desktop_lifecycle_plugin.h>
+#include <local_notifier/local_notifier_plugin.h>
 #include <protocol_handler/protocol_handler_plugin.h>
 #include <proxy_manager/proxy_manager_plugin.h>
 #include <screen_retriever/screen_retriever_plugin.h>
 #include <tray_manager/tray_manager_plugin.h>
-#include <vclibs/vclibs_plugin.h>
+#include <url_launcher_windows/url_launcher_windows.h>
 #include <window_manager/window_manager_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  DesktopLifecyclePluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("DesktopLifecyclePlugin"));
+  LocalNotifierPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("LocalNotifierPlugin"));
   ProtocolHandlerPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("ProtocolHandlerPlugin"));
   ProxyManagerPluginRegisterWithRegistrar(
@@ -22,8 +28,8 @@ void RegisterPlugins(flutter::PluginRegistry* registry) {
       registry->GetRegistrarForPlugin("ScreenRetrieverPlugin"));
   TrayManagerPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("TrayManagerPlugin"));
-  VclibsPluginRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("VclibsPlugin"));
+  UrlLauncherWindowsRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("UrlLauncherWindows"));
   WindowManagerPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("WindowManagerPlugin"));
 }
